Rework a lot of things. Authenticating is supported.

+++ app.rb
@@ -1,84 +1,58 @@
 require 'sinatra/base'
-require 'httparty'
+require './github_party'
 
-def httparty_error(r)
-  "<a href='#{r.request.path.to_s}'>#{r.request.path.to_s}</a>: #{r.code} #{r.message}: #{r.body}. #{r.headers.to_h.to_json}"
-end
+app_path = File.expand_path(File.dirname(__FILE__))
+Dir["#{app_path}/config/initializers/*.rb"].each { |f| require f }
 
-def get_gists(user)
-  redis_data = $redis.get 'gists'
-  comments = []
-  if redis_data
-    entries = JSON.parse redis_data
-  else
-    page = 1
-    entries = []
-    while true
-      r = HTTParty.get "https://api.github.com/users/#{user}/gists?page=#{page}"
-      return httparty_error(r) unless r.success?
-      http_json = JSON.parse r.body
-      break if http_json.count == 0
-      entries = entries + http_json
-      page += 1
-    end
-    $redis.set 'gists', entries.to_json
-    $redis.expire 'gists', 60*60
-  end
-  entries
-end
-
-def get_comments(gists)
-  gists.map do |gist|
-    id = gist['id']
-    redis_data = $redis.get "gist:#{id}"
-    if redis_data
-      comments = JSON.parse redis_data
-    else
-      r = HTTParty.get gist['comments_url']
-      return httparty_error(r) unless r.success?
-      comments = JSON.parse r.body
-      $redis.set "gist:#{id}", comments.to_json
-      $redis.expire "gist:#{id}", 60*30
-    end
-
-    comments.map do |c|
-      [ id, c ]
-    end
-  end.flatten(1).sort_by { |id, c| c['created_at'] }.reverse!
+def format_date(date)
+  date.gsub('T',' ').gsub('Z','')
 end
 
 class App < Sinatra::Base
   get '/' do
-    "<a href='/gists/stefansundin.xml'>stefansundin</a> <a href='/flushall'>flush cache</a>"
+    "<h1>GitHub activity rss feed</h1>"+
+    "<p><form action='/go' method='get'>username: <input type='text' name='username' placeholder='github username' value='stefansundin'> <input type='submit' value='Go'></form></p>"+
+    "<p><a href='/stefansundin.xml'>stefansundin</a> <a href='/flush'>flush cache</a></p>"+
+    "<p>authenticated? "+($redis.exists('access_token') ? $redis.get('login') : "<a href='/auth'>no</a></p>") +
+    "<p>ratelimit: #{GithubParty.ratelimit.to_json}</p>"
   end
 
-  get '/gists/:user.xml' do
+  get '/go' do
+    redirect "/#{params[:username]}.xml"
+  end
+
+  get '/:user.xml' do
+    client = GithubParty.new
     user = params[:user]
-    gists = get_gists user
-    #return gists.to_json
-    return "Error: #{gists}" if gists.is_a? String
-    comments = get_comments gists
-    #return comments.to_json
-    return "Error: #{comments}" if comments.is_a? String
-    comments.delete_if { |id, c| c['user']['login'] == user }
+    gists = client.gists user
+
+    gist_comments = gists.map do |gist|
+      comments = client.gist_comments gist
+      comments.map do |comment|
+        [ gist['id'], comment ]
+      end
+    end.flatten(1)
+    gist_comments.reject! { |id, c| c['user']['login'] == user }
+    gist_comments.sort_by! { |id, c| c['created_at'] }
+    gist_comments.reverse!
 
     headers 'Content-Type' => 'application/atom+xml'
     output = []
     output.push <<eos
 <?xml version="1.0" encoding="utf-8"?>
 <feed xmlns="http://www.w3.org/2005/Atom">
-  <id>gist:#{user}</id>
+  <id>user:#{user}</id>
   <title>#{user} gist comments</title>
   <icon>https://assets-cdn.github.com/favicon.ico</icon>
-  <link href="http://stefansundin.com:3000/gists/#{user}" rel="self" />
-  <updated>#{comments[0][1]['updated_at'] if comments[0]}</updated>
+  <link href="http://stefansundin.com:3000/#{user}.xml" rel="self" />
+  <updated>#{gist_comments[0][1]['updated_at'] if gist_comments[0]}</updated>
   <author>
     <name>Stefan Sundin</name>
     <uri>https://stefansundin.com/</uri>
   </author>
 eos
 
-    comments.map do |id, c|
+    gist_comments.map do |id, c|
       output.push <<eos
 
   <entry>
@@ -88,7 +62,7 @@ eos
     <updated>#{c['updated_at']}</updated>
     <author><name>#{c['user']['login']}</name></author>
     <content type="html">
-&lt;p>User #{c['user']['login']} commented on #{id} at #{c['created_at']}:&lt;/p>
+&lt;p>User #{c['user']['login']} commented on #{id} at #{format_date(c['created_at'])}:&lt;/p>
 &lt;pre>
 #{c['body']}
 &lt;/pre>
@@ -101,8 +75,30 @@ eos
     output.join('')
   end
 
+  get '/flush' do
+    $redis.keys.each do |key|
+      if not %w[access_token login].include?(key)
+        $redis.del key
+      end
+    end
+    redirect '/'
+  end
+
   get '/flushall' do
+    if ENV['RACK_ENV'] == 'production'
+      return 'forbidden in production'
+    end
     $redis.flushall
+    redirect '/'
+  end
+
+  get '/auth' do
+    redirect "https://github.com/login/oauth/authorize?client_id=#{ENV['GITHUB_CLIENT_ID']}"
+  end
+
+  get '/callback' do
+    return 'already authenticated' if $redis.exists('access_token')
+    GithubParty.authenticate(request.env['rack.request.query_hash']['code'])
+    redirect '/'
   end
 end
-
